+++ payments/RELEASES.md
@@ -88,3 +88,13 @@ Based on:
 - [typescript v2.2.0] payments
 ### Releases
 - [NPM v2.2.0] https://www.npmjs.com/package/@wingspan/payments/v/2.2.0 - payments
+
+## 2024-08-12 01:31:42
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0-oas3.1 ./openapi/payment_speakeasy_edit.yaml
+- Speakeasy CLI 1.361.1 (2.393.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.3.0] payments
+### Releases
+- [NPM v2.3.0] https://www.npmjs.com/package/@wingspan/payments/v/2.3.0 - payments